% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descdist2.R
\name{descdist2}
\alias{descdist2}
\title{Descdist but with colors}
\usage{
descdist2(
  data,
  discrete = FALSE,
  boot = NULL,
  method = "unbiased",
  graph = TRUE,
  print = TRUE,
  obs.col = "red",
  obs.pch = 16,
  boot.col = "orange",
  data.col = "lightgrey"
)
}
\description{
Adds bg/data color to descdist plot
}
