% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare different machine learning models}
\usage{
compare(train_data, test_data, caret_method)
}
\arguments{
\item{train_data}{a list containing one or multiple data frames}

\item{test_data}{a single data frame}

\item{caret_methods}{a string with one or vector with multiple caret's train methods}
}
\description{
Trains a machine learning model on a single or different
train data frames using caret train with given methods,
tests the model on a test data frames,
and returns a data frame with names of passed parameters
and results of model training
}
